Solved task 102 <2>

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	long long i = 0, prev = 1, next = 1, current;
+	long i = 0, prev = 1, next = 1, current;
 
 	/*1, 1, 2, 3, 5, 8,...*/
 	for (; i < 48; i++)
@@ -14,7 +14,7 @@ int main(void)
 		current = next + prev;
 		prev = next;
 		next = current;
-		printf("%lld, ", prev);
+		printf("%ld, ", prev);
 
 	}
 	putchar('\n');
